pyproject.toml
[project]
name = "stock-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4==4.14.3",
    "certifi==2026.1.4",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "contourpy==1.3.3",
    "curl-cffi==0.13.0",
    "cycler==0.12.1",
    "defusedxml==0.7.1",
    "fonttools==4.61.1",
    "fpdf2==2.8.5",
    "frozendict==2.4.7",
    "idna==3.11",
    "kiwisolver==1.4.9",
    "matplotlib==3.10.8",
    "multitasking==0.0.12",
    "numpy==2.4.1",
    "open-tamil==1.1",
    "packaging==26.0",
    "pandas==3.0.0",
    "peewee==3.19.0",
    "pillow==12.1.0",
    "platformdirs==4.5.1",
    "protobuf==6.33.4",
    "pycparser==3.0",
    "pyparsing==3.3.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "pytz==2025.2",
    "requests==2.32.5",
    "scipy==1.17.0",
    "seaborn==0.13.2",
    "six==1.17.0",
    "soupsieve==2.8.3",
    "tamil==0.3",
    "typing-extensions==4.15.0",
    "urllib3==2.6.3",
    "websockets==16.0",
    "yfinance==1.1.0",
]
